refactor: Change packages of command (#1651)
