+++ test_runner/src/main/kotlin/ftl/cli/firebase/test/AndroidCommand.kt
@@ -1,12 +1,12 @@
 package ftl.cli.firebase.test
 
 import ftl.cli.firebase.test.android.AndroidDoctorCommand
+import ftl.cli.firebase.test.android.AndroidLocalesCommand
+import ftl.cli.firebase.test.android.AndroidModelsCommand
+import ftl.cli.firebase.test.android.AndroidOrientationsCommand
 import ftl.cli.firebase.test.android.AndroidRunCommand
 import ftl.cli.firebase.test.android.AndroidTestEnvironmentCommand
-import ftl.cli.firebase.test.android.configuration.AndroidLocalesCommand
-import ftl.cli.firebase.test.android.models.AndroidModelsCommand
-import ftl.cli.firebase.test.android.orientations.AndroidOrientationsCommand
-import ftl.cli.firebase.test.android.versions.AndroidVersionsCommand
+import ftl.cli.firebase.test.android.AndroidVersionsCommand
 import picocli.CommandLine
 import picocli.CommandLine.Command
