+++ test_runner/src/main/kotlin/ftl/cli/firebase/test/IosCommand.kt
@@ -1,12 +1,12 @@
 package ftl.cli.firebase.test
 
 import ftl.cli.firebase.test.ios.IosDoctorCommand
+import ftl.cli.firebase.test.ios.IosLocalesCommand
+import ftl.cli.firebase.test.ios.IosModelsCommand
+import ftl.cli.firebase.test.ios.IosOrientationsCommand
 import ftl.cli.firebase.test.ios.IosRunCommand
 import ftl.cli.firebase.test.ios.IosTestEnvironmentCommand
-import ftl.cli.firebase.test.ios.configuration.IosLocalesCommand
-import ftl.cli.firebase.test.ios.models.IosModelsCommand
-import ftl.cli.firebase.test.ios.orientations.IosOrientationsCommand
-import ftl.cli.firebase.test.ios.versions.IosVersionsCommand
+import ftl.cli.firebase.test.ios.IosVersionsCommand
 import picocli.CommandLine
 import picocli.CommandLine.Command
