+++ test_runner/src/main/kotlin/ftl/cli/firebase/test/android/AndroidOrientationsCommand.kt
@@ -1,5 +1,6 @@
-package ftl.cli.firebase.test.ios.orientations
+package ftl.cli.firebase.test.android
 
+import ftl.cli.firebase.test.android.orientations.AndroidOrientationsListCommand
 import picocli.CommandLine
 
 @CommandLine.Command(
@@ -9,13 +10,13 @@ import picocli.CommandLine
     descriptionHeading = "%n@|bold,underline Description:|@%n%n",
     parameterListHeading = "%n@|bold,underline Parameters:|@%n",
     optionListHeading = "%n@|bold,underline Options:|@%n",
-    header = ["Information about available orientation versions"],
+    header = ["Information about available orientations"],
     description = ["Prints list of available orientations"],
-    subcommands = [IosOrientationsListCommand::class],
+    subcommands = [AndroidOrientationsListCommand::class],
     usageHelpAutoWidth = true
 )
-class IosOrientationsCommand : Runnable {
+class AndroidOrientationsCommand : Runnable {
     override fun run() {
-        CommandLine.usage(IosOrientationsCommand(), System.out)
+        CommandLine.usage(AndroidOrientationsCommand(), System.out)
     }
 }
