+++ test_runner/src/main/kotlin/ftl/cli/firebase/test/android/AndroidVersionsCommand.kt
@@ -1,21 +1,23 @@
-package ftl.cli.firebase.test.ios.versions
+package ftl.cli.firebase.test.android
 
+import ftl.cli.firebase.test.android.versions.AndroidVersionsDescribeCommand
+import ftl.cli.firebase.test.android.versions.AndroidVersionsListCommand
 import picocli.CommandLine
 
 @CommandLine.Command(
     name = "versions",
-    subcommands = [IosVersionsListCommand::class, IosVersionsDescribeCommand::class],
     headerHeading = "",
     synopsisHeading = "%n",
     descriptionHeading = "%n@|bold,underline Description:|@%n%n",
     parameterListHeading = "%n@|bold,underline Parameters:|@%n",
     optionListHeading = "%n@|bold,underline Options:|@%n",
-    header = ["Describe software versions"],
+    header = ["Information about available software versions"],
     description = ["Information about available software versions. For example prints list of available software versions"],
+    subcommands = [AndroidVersionsListCommand::class, AndroidVersionsDescribeCommand::class],
     usageHelpAutoWidth = true
 )
-class IosVersionsCommand : Runnable {
+class AndroidVersionsCommand : Runnable {
     override fun run() {
-        CommandLine.usage(IosVersionsCommand(), System.out)
+        CommandLine.usage(AndroidVersionsCommand(), System.out)
     }
 }
