+++ test_runner/src/main/kotlin/ftl/cli/firebase/test/ios/IosVersionsCommand.kt
@@ -1,21 +1,23 @@
-package ftl.cli.firebase.test.android.versions
+package ftl.cli.firebase.test.ios
 
+import ftl.cli.firebase.test.ios.versions.IosVersionsDescribeCommand
+import ftl.cli.firebase.test.ios.versions.IosVersionsListCommand
 import picocli.CommandLine
 
 @CommandLine.Command(
     name = "versions",
+    subcommands = [IosVersionsListCommand::class, IosVersionsDescribeCommand::class],
     headerHeading = "",
     synopsisHeading = "%n",
     descriptionHeading = "%n@|bold,underline Description:|@%n%n",
     parameterListHeading = "%n@|bold,underline Parameters:|@%n",
     optionListHeading = "%n@|bold,underline Options:|@%n",
-    header = ["Information about available software versions"],
+    header = ["Describe software versions"],
     description = ["Information about available software versions. For example prints list of available software versions"],
-    subcommands = [AndroidVersionsListCommand::class, AndroidVersionsDescribeCommand::class],
     usageHelpAutoWidth = true
 )
-class AndroidVersionsCommand : Runnable {
+class IosVersionsCommand : Runnable {
     override fun run() {
-        CommandLine.usage(AndroidVersionsCommand(), System.out)
+        CommandLine.usage(IosVersionsCommand(), System.out)
     }
 }
